Rename v10 modules
