+++ src/prelude.rs
@@ -1,4 +1,4 @@
 //! Convenience wrapper to import all of Hypercore's core.
 pub use crate::common::Store;
 pub use crate::core::Hypercore;
-pub use crate::storage_v10::{PartialKeypair, Storage};
+pub use crate::storage::{PartialKeypair, Storage};
